+++ packages/did-ui-react/src/utils/socialLoginAuth.ts
@@ -1,10 +1,17 @@
-import { getCustomNetworkType, getServiceUrl } from '../components/config-provider/utils';
-import { PORTKEY_VERSION, WEB_PAGE, WEB_PAGE_TEST } from '../constants';
+import {
+  getCustomNetworkType,
+  getServiceUrl,
+  getSocketUrl,
+  getStorageInstance,
+} from '../components/config-provider/utils';
+import { WEB_PAGE, WEB_PAGE_TEST } from '../constants';
 import { ISocialLogin, NetworkType } from '../types';
 import { stringify } from 'query-string';
 import { dealURLLastChar } from './lib';
+import OpenLogin from './openlogin';
+import { facebookAuthPath, twitterAuthPath } from './openlogin/contants';
 
-export const socialLoginAuth = ({
+export const socialLoginAuthOpener = ({
   type,
   clientId,
   redirectURI,
@@ -23,11 +30,37 @@ export const socialLoginAuth = ({
   new Promise((resolve, reject) => {
     let timer: any = null;
     let serviceURI = dealURLLastChar(serviceUrl);
+    let _redirectURI = redirectURI;
 
-    if (type === 'Telegram' && !serviceURI) serviceURI = getServiceUrl();
+    if ((type === 'Telegram' || type === 'Facebook' || type === 'Twitter') && !serviceURI) serviceURI = getServiceUrl();
+    if (!redirectURI) {
+      switch (type) {
+        case 'Facebook':
+          _redirectURI = `${serviceURI}${facebookAuthPath}`;
+          break;
+        case 'Twitter':
+          _redirectURI = `${serviceURI}${twitterAuthPath}`;
+          break;
+      }
+    }
+
     const ctw = getCustomNetworkType();
 
-    const thirdPage = ctw === 'Offline' ? WEB_PAGE_TEST : WEB_PAGE;
+    let thirdPage;
+    switch (ctw) {
+      case 'Offline':
+        thirdPage = WEB_PAGE_TEST;
+        break;
+      case 'onLine':
+        thirdPage = WEB_PAGE;
+        break;
+
+      case 'local':
+        thirdPage = 'http://localhost:3000';
+        break;
+      default:
+        thirdPage = WEB_PAGE;
+    }
 
     const onMessage = (event: MessageEvent) => {
       const type = event.data.type;
@@ -49,6 +82,7 @@ export const socialLoginAuth = ({
 
     window.addEventListener('message', onMessage);
     const baseUrl = `${thirdPage}/social-login/${type}`;
+    const encode = !(type === 'Twitter' || type === 'Facebook');
     const queryParams =
       type === 'Telegram'
         ? {
@@ -56,8 +90,14 @@ export const socialLoginAuth = ({
             from: 'openlogin',
             serviceURI,
           }
-        : { clientId, redirectURI, version: PORTKEY_VERSION };
-    const windowOpener = window.open(`${baseUrl}?${stringify(queryParams)}`);
+        : {
+            clientId,
+            redirectURI: _redirectURI,
+            // version: PORTKEY_VERSION
+          };
+
+    console.log(`${baseUrl}?${stringify(queryParams, { encode })}`, '=====baseUrl');
+    const windowOpener = window.open(`${baseUrl}?${stringify(queryParams, { encode })}`);
 
     timer = setInterval(() => {
       if (windowOpener?.closed) {
@@ -67,3 +107,40 @@ export const socialLoginAuth = ({
       }
     }, 1600);
   });
+
+export const socialLoginAuthBySocket = async ({
+  type,
+  clientId,
+}: {
+  type: ISocialLogin;
+  clientId?: string;
+  redirectURI?: string;
+  network?: NetworkType;
+  serviceUrl?: string;
+}): Promise<{
+  token: string;
+  provider: ISocialLogin;
+}> => {
+  const serviceURI = getServiceUrl();
+  const socketURI = getSocketUrl();
+
+  const openlogin = new OpenLogin({
+    network: getCustomNetworkType(),
+    serviceURI: serviceURI,
+    clientId,
+    socketURI,
+    currentStorage: getStorageInstance(),
+    sdkUrl: 'http://localhost:3000',
+  });
+
+  const result = await openlogin.login({
+    from: 'openlogin',
+    loginProvider: type,
+  });
+  if (!result) throw 'Not result';
+  if (result?.code) throw result.message;
+  console.log(result, 'result===');
+  return result;
+};
+
+export const socialLoginAuth = socialLoginAuthBySocket;