+++ packages/next-example/rewrites.js
@@ -14,16 +14,16 @@ module.exports = [
     // { source: '/graphql/:path*', destination: 'http://192.168.64.202:8083/AElfIndexer_DApp/PortKeyIndexerCASchema/graphql/:path*' },
 
     // // test4
-    // { source: '/api/:path*', destination: 'http://192.168.67.179:5001/api/:path*' },
-    // { source: '/connect/:path*', destination: 'http://192.168.67.179:8001/connect/:path*' },
-    // { source: '/graphql/:path*', destination: 'http://192.168.67.67:8083/AElfIndexer_DApp/PortKeyIndexerCASchema/graphql/:path*' },
+    { source: '/api/:path*', destination: 'http://192.168.66.117:5577/api/:path*' },
+    { source: '/connect/:path*', destination: 'http://192.168.66.117:8080/connect/:path*' },
+    { source: '/graphql/:path*', destination: 'http://192.168.67.214:8083/AElfIndexer_DApp/PortKeyIndexerCASchema/graphql/:path*' },
 
     // mainnet
     // { source: '/api/:path*', destination: 'https://did-portkey.portkey.finance/api/:path*' },
     // { source: '/connect/:path*', destination: 'https://auth-portkey.portkey.finance/connect/:path*' },
 
     // testnet
-    { source: '/api/:path*', destination: 'https://aa-portkey-test.portkey.finance/api/:path*' },
-    { source: '/connect/:path*', destination: 'https://auth-aa-portkey-test.portkey.finance/connect/:path*' },
+    // { source: '/api/:path*', destination: 'https://aa-portkey-test.portkey.finance/api/:path*' },
+    // { source: '/connect/:path*', destination: 'https://auth-aa-portkey-test.portkey.finance/connect/:path*' },
 
 ];